Fixed compile warnings under linux.
git-svn-id: svn://localhost:13690/uGuru/trunk@21 3018da27-9c45-1242-9092-7a69cc569eec

+++ uGuru/include/datetime.h
@@ -12,9 +12,9 @@ const char * mg_getdate();
 void mg_wait(int secs);
 
 /*
- *	Calculates the difference in seconds between t1 and t2.
+ *    Calculates the difference in seconds between t1 and t2.
  */
-long mg_difftime(long t1, long t2);
+double mg_difftime(long t1, long t2);
 
 #ifdef __cplusplus
 }
